requirements.txt
# ALFRED_UBX Requirements
# AI Assistant with Persistent Memory & Adaptive Learning
# Author: Daniel J. Rita aka BATDAN007
# https://github.com/Batdan007/ALFRED_UBX

# ===== AI Providers =====
anthropic>=0.39.0
openai>=1.50.0
groq>=0.11.0

# ===== Web Framework =====
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
starlette>=0.41.0

# ===== MCP (Model Context Protocol) =====
mcp>=1.0.0

# ===== CLI & UI =====
rich>=13.9.0
prompt-toolkit>=3.0.48
typer>=0.13.0

# ===== Text-to-Speech =====
pyttsx3>=2.90

# ===== Data & Config =====
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1

# ===== HTTP Clients =====
aiohttp>=3.10.0
httpx>=0.27.0
requests>=2.32.0

# ===== Database & Storage =====
aiosqlite>=0.20.0
sqlalchemy>=2.0.0

# ===== Async Support =====
anyio>=4.6.0
asyncio-atexit>=1.0.1

# ===== Utilities =====
python-dateutil>=2.9.0
tenacity>=9.0.0
PyYAML>=6.0.2

# ===== Testing (optional) =====
pytest>=8.3.0
pytest-cov>=5.0.0
pytest-asyncio>=0.24.0

# ===== Development (optional) =====
black>=24.0.0
ruff>=0.7.0
mypy>=1.13.0

# ===== Platform-Specific =====
# Windows: pywin32 (installed automatically by pyttsx3)
# macOS: pyobjc-framework-Cocoa (for speech)
# Linux: espeak-ng system package required
